requirements.txt
asgiref==3.3.4
beautifulsoup4==4.10.0
black==21.12b0
click==8.0.3
Django==3.2
django-ckeditor==6.0.0
django-cleanup==6.0.0
django-crispy-forms==1.11.2
django-htmlmin==0.11.0
django-js-asset==1.2.2
html5lib==1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
python-dotenv==0.19.2
pytz==2021.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
tomli==1.2.3
typing-extensions==4.0.1
webencodings==0.5.1
